Added a perfect material.
I create a material called Perfect_M that a material instance can be
created from(right click, first option).

This is "perfect" because it can be applied to any object, at which
point parameters can be changed to better suite you object. The default
texture is crate.

Additionally, I added distance based tessellation to Perfect_M. This is
expensive to implement, but not the worst case. The alpha or
displacement map of the object is used with tessellation to create
realistic geometry, but only when the player is near the object. It is
reduced as they walk away. The values for distance can be changed in the
material instance details menu. Example: Stone wall texture, the stones
become geometry.